+++ db/access.php
@@ -37,6 +37,18 @@ $capabilities = array(
         ),
     ),
 
+    'local/ildmeta:delete_sponsor' => array(
+        'riskbitmask' => RISK_SPAM | RISK_XSS,
+        'captype' => 'write',
+        'contextlevel' => CONTEXT_COURSE,
+        'archetypes' => array(
+            'manager' => CAP_ALLOW,
+            'editingteacher' => CAP_ALLOW,
+            #'kursmanager' => CAP_ALLOW,
+            'coursecreator' => CAP_ALLOW
+        ),
+    ),
+
     'local/ildmeta:indexation' => array(
         'riskbitmask' => RISK_SPAM | RISK_XSS,
         'captype' => 'write',